+++ spec/punchblock/translator/asterisk_spec.rb
@@ -151,6 +151,13 @@ module Punchblock
           end
         end
 
+        let :end_error_event do
+          Punchblock::Event::End.new.tap do |e|
+            e.target_call_id = call_id
+            e.reason = :error
+          end
+        end
+
         context "for an outgoing call which began executing but crashed" do
           let(:dial_command) { Command::Dial.new :to => 'SIP/1234', :from => 'abc123' }
 
@@ -167,6 +174,8 @@ module Punchblock
             call.wrapped_object.define_singleton_method(:oops) do
               raise 'Woops, I died'
             end
+
+            connection.expects(:handle_event).once.with end_error_event
 
             lambda { call.oops }.should raise_error(/Woops, I died/)
             sleep 0.1
@@ -192,7 +201,7 @@ module Punchblock
           let(:call_id) { call.id }
 
           before do
-            subject.wrapped_object.stubs :handle_pb_event
+            connection.expects(:handle_event).at_least(1)
             subject.handle_ami_event ami_event
             call_id
           end
@@ -201,6 +210,8 @@ module Punchblock
             call.wrapped_object.define_singleton_method(:oops) do
               raise 'Woops, I died'
             end
+
+            connection.expects(:handle_event).once.with end_error_event
 
             lambda { call.oops }.should raise_error(/Woops, I died/)
             sleep 0.1