+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * Bugfix: Asterisk calls and components are removed from registries when they die
   * Bugfix: Commands for unknown calls/components respond with the correct `:item_not_found` name
   * Bugfix: Asterisk calls send an error complete event for their dying components
+  * Bugfix: Asterisk translator sends an error end event for its dying calls
 
 # [v1.2.0](https://github.com/adhearsion/punchblock/compare/v1.1.0...v1.2.0) - [2012-04-29](https://rubygems.org/gems/punchblock/versions/1.2.0)
   * Feature: Basic support for record component on Asterisk, using MixMonitor. Currently unsupported options include: start_paused, initial_timeout, final_timeout. Hints are additionally not supported, and recordings are stored on the * machine's local filesystem.